Fixed crash reason. Now dll can be used only with C++ wrapper

+++ MinecraftESP/Structs/Options/Tracer.cs
@@ -5,4 +5,4 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace ESP.Structs.Options;
-public record struct Tracer(bool Enabled, Color Color, int LineWidth, float OffsetX = 0, float OffsetY = 0, float OffsetZ = 0);
+public record struct Tracer(bool Enabled, Color Color, float LineWidth, float OffsetX = 0, float OffsetY = 0, float OffsetZ = 0);
